refactor(dashboard): extract Mastio audit sub-router (F-B-201 PR-7/10)
Move the /proxy/audit route + the private _pretty_and_recipient helper
out of the ~3500-LOC mcp_proxy/dashboard/router.py god-object into a
dedicated audit_routes.py module, mounted via include_router.

Same pattern as PR-2..PR-6. The handler keeps both streams (admin
audit_log + hash-chained local_audit) merged into a single unified
view exactly as before.

F-B-201 progress: 7/10 PR.
